gracefully handled pipes in existing cases
https://users.rust-lang.org/t/why-does-the-pipe-cause-the-panic-in-the-standard-library/107222/4
https://github.com/rust-lang/rust/issues/97889

unfortunately the appropriate rustlang feature is still experimental;
in reality, the most convenient handling should be done by rustlang
itself, since this is a fragile case-by-case basis.

in order for this approach to be correct throughout, all `println!`
macros should be replaced with the appropriate `writeln!` handling;
admittedly, this is too invasive for such a minor issue.