implemented generic soutln macro
this macro should behave exactly as `println!` would, except for
specific error handling, which println! does not.
there is a case we do not handle yet, which I do not know how to handle
or how to make it occur, where println would fail with non-blocking
stdout, however i am not even sure if it is applicable here.

according to the docstrings, the following might be implementation
dependent, however it looks like soutln! (via writeln!) will also
flush to stdout on each call since `stdout` is line-buffered.

`soutln!` syntax conforms to the same syntax as `println!` and
should be mass-replaced in the future, when the appropriate
rustlang merge has been completed